+++ POSHOrigin/POSHOrigin.psm1
@@ -20,6 +20,13 @@ $moduleRoot = Split-Path -Path $MyInvocation.MyCommand.Path
     Where-Object { -not ($_.ProviderPath.ToLower().Contains(".tests.")) } |
     ForEach-Object { . $_.ProviderPath }
 
+# This gets executing when the module is removed from the session
+$MyInvocation.MyCommand.ScriptBlock.Module.OnRemove = {
+    Remove-Variable -Name credentialCache -Scope Script
+    Remove-Variable -Name modulesToProcess -Scope Script
+    Remove-Variable -Name resourceCache -Scope Script         
+}  
+
 #Export-ModuleMember -Function *POSH*
 New-Alias -Name gpoc -Value Get-POSHOriginConfig
 New-Alias -Name secret -Value Get-POSHOriginSecret
@@ -29,7 +36,6 @@ New-Alias -Name ipo -Value Invoke-POSHOrigin
 New-Alias -Name iponew -Value Invoke-POSHOriginNEW
 New-Alias -Name resource -Value New-POSHOriginResource
 New-Alias -Name module -Value New-POSHOriginResourceFromModule
-New-Alias -Name wps -Value Write-POSHScreen
 #Export-ModuleMember -Alias gpoc
 #Export-ModuleMember -Alias gpos
 #Export-ModuleMember -Alias gpd