+++ psgallery.psdeploy.ps1
@@ -1,0 +1,9 @@
+Deploy POSHOrigin {
+    By PSGalleryModule ToPSGallery {
+        FromSource '.\POSHOrigin'
+        To 'PSGallery'
+        WithOptions @{
+            ApiKey = $env:PSGALLERY_API_KEY
+        }
+    }
+}